Fixed sources and UTs so that they compile and run without errors. Fixed CMake config
